requirements.txt
Flask>=3.0.0
requests>=2.32.0
pyjwt>=2.8.0
python-dotenv>=1.0.1